Refactor auth flow
