+++ packages/react/src/components/form/org-switching/useOrgSwitchingFlowState.ts
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useRef, useState } from "react";
 
 import { useSlashID } from "../../../hooks/use-slash-id";
-import { Flow, createFlow } from "./org-switching-flow";
+import { Flow, createFlow } from "../flow/org-switching-flow";
 import { Handle } from "../../../domain/types";
 import { FlowState, CreateFlowOptions } from "../flow/flow.common";
